refactor[api-server]: lottery.repository 패키지 위치 변경
