+++ src/main/java/com/watermelon/server/event/lottery/repository/LotteryRewardRepository.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.lottery.repository;
+package com.watermelon.server.event.lottery.repository;
 
 import com.watermelon.server.event.lottery.domain.LotteryReward;
 import org.springframework.data.jpa.repository.JpaRepository;
