+++ src/test/java/com/watermelon/server/lottery/repository/LotteryApplierRepositoryTest.java
@@ -2,6 +2,7 @@ package com.watermelon.server.lottery.repository;
 
 import com.watermelon.server.event.lottery.domain.Link;
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
+import com.watermelon.server.event.lottery.repository.LotteryApplierRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
