+++ src/test/java/com/watermelon/server/lottery/service/ExpectationServiceTest.java
@@ -4,7 +4,7 @@ import com.watermelon.server.event.lottery.domain.Expectation;
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
 import com.watermelon.server.event.lottery.dto.request.RequestExpectationDto;
 import com.watermelon.server.event.lottery.error.ExpectationAlreadyExistError;
-import com.watermelon.server.lottery.repository.ExpectationRepository;
+import com.watermelon.server.event.lottery.repository.ExpectationRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
